Split pytest into separate files and add factory for detectors
